ref(crons): Show processing errors on cron detectors list
Add processing errors alert to the cron detectors list page to match
the existing behavior on the insights crons overview page. Created a
shared GlobalMonitorProcessingErrors component to avoid code duplication.

Fixes [NEW-633](https://linear.app/getsentry/issue/NEW-633/general-cron-errors-should-be-shown-on-the-cron-detectors-specific)
