+++ static/app/views/insights/crons/views/overview.tsx
@@ -5,7 +5,6 @@ import * as qs from 'query-string';
 import {Alert} from '@sentry/scraps/alert';
 
 import {openBulkEditMonitorsModal} from 'sentry/actionCreators/modal';
-import {deleteProjectProcessingErrorByType} from 'sentry/actionCreators/monitors';
 import {Button} from 'sentry/components/core/button';
 import {ButtonBar} from 'sentry/components/core/button/buttonBar';
 import FeedbackWidgetButton from 'sentry/components/feedback/widget/feedbackWidgetButton';
@@ -30,7 +29,6 @@ import {useApiQuery} from 'sentry/utils/queryClient';
 import {decodeList, decodeScalar} from 'sentry/utils/queryString';
 import useRouteAnalyticsEventNames from 'sentry/utils/routeAnalytics/useRouteAnalyticsEventNames';
 import useRouteAnalyticsParams from 'sentry/utils/routeAnalytics/useRouteAnalyticsParams';
-import useApi from 'sentry/utils/useApi';
 import {useLocation} from 'sentry/utils/useLocation';
 import {useNavigate} from 'sentry/utils/useNavigate';
 import useOrganization from 'sentry/utils/useOrganization';
@@ -42,14 +40,9 @@ import {
 import {NewMonitorButton} from 'sentry/views/insights/crons/components/newMonitorButton';
 import {OverviewTimeline} from 'sentry/views/insights/crons/components/overviewTimeline';
 import {OwnerFilter} from 'sentry/views/insights/crons/components/ownerFilter';
-import {MonitorProcessingErrors} from 'sentry/views/insights/crons/components/processingErrors/monitorProcessingErrors';
-import {makeMonitorListErrorsQueryKey} from 'sentry/views/insights/crons/components/processingErrors/utils';
+import {GlobalMonitorProcessingErrors} from 'sentry/views/insights/crons/components/processingErrors/globalMonitorProcessingErrors';
 import {MODULE_DESCRIPTION, MODULE_DOC_LINK} from 'sentry/views/insights/crons/settings';
-import type {
-  CheckinProcessingError,
-  Monitor,
-  ProcessingErrorType,
-} from 'sentry/views/insights/crons/types';
+import type {Monitor} from 'sentry/views/insights/crons/types';
 import {makeMonitorListQueryKey} from 'sentry/views/insights/crons/utils';
 
 const CronsListPageHeader = HookOrDefault({
@@ -57,7 +50,6 @@ const CronsListPageHeader = HookOrDefault({
 });
 
 function CronsOverview() {
-  const api = useApi();
   const organization = useOrganization();
   const navigate = useNavigate();
   const location = useLocation();
@@ -76,13 +68,6 @@ function CronsOverview() {
     staleTime: 0,
   });
 
-  const processingErrorQueryKey = makeMonitorListErrorsQueryKey(organization, project);
-  const {data: processingErrors, refetch: refetchErrors} = useApiQuery<
-    CheckinProcessingError[]
-  >(processingErrorQueryKey, {
-    staleTime: 0,
-  });
-
   useRouteAnalyticsEventNames('monitors.page_viewed', 'Monitors: Page Viewed');
   useRouteAnalyticsParams({empty_state: !monitorList || monitorList.length === 0});
 
@@ -95,16 +80,6 @@ function CronsOverview() {
       query: normalizeDateTimeParams({...currentQuery, query}),
     });
   };
-
-  async function handleDismissError(errorType: ProcessingErrorType, projectId: string) {
-    await deleteProjectProcessingErrorByType(
-      api,
-      organization.slug,
-      projectId,
-      errorType
-    );
-    await refetchErrors();
-  }
 
   const showAddMonitor = !isValidPlatform(platform) || !isValidGuide(guide);
 
@@ -171,18 +146,9 @@ function CronsOverview() {
               onSearch={handleSearch}
             />
           </Filters>
-          {!!processingErrors?.length && (
-            <Alert.Container>
-              <MonitorProcessingErrors
-                checkinErrors={processingErrors}
-                onDismiss={handleDismissError}
-              >
-                {t(
-                  'Errors were encountered while ingesting check-ins for the selected projects'
-                )}
-              </MonitorProcessingErrors>
-            </Alert.Container>
-          )}
+          <Alert.Container>
+            <GlobalMonitorProcessingErrors project={project} />
+          </Alert.Container>
           {isPending ? (
             <LoadingIndicator />
           ) : monitorList?.length ? (
